app/requirements.txt
flask
psycopg2-binary
gunicorn
google-cloud-storage

